+++ tests/run.lisp
@@ -16,6 +16,7 @@
 (load (project-file "src/application/arm-run.lisp"))
 (load (project-file "src/application/start-run.lisp"))
 (load (project-file "src/application/checkpoint.lisp"))
+(load (project-file "src/application/close-run.lisp"))
 
 (in-package :hngh.tests)
 
@@ -41,6 +42,7 @@
 (load (cl-user::project-file "tests/application/test-arm-run.lisp"))
 (load (cl-user::project-file "tests/application/test-start-run.lisp"))
 (load (cl-user::project-file "tests/application/test-checkpoint.lisp"))
+(load (cl-user::project-file "tests/application/test-close-run.lisp"))
 
 (defun read-fixture (relative)
   (with-open-file (stream (cl-user::project-file relative))